move run examples into subpackage
